chore(tests): :truck: Move test from tests to src/tests
