+++ src/tests/utils/response.test.ts
@@ -1,8 +1,8 @@
 import {
-  describe, expect, it,
+    describe, expect, it,
 } from '@jest/globals';
 import { mockResponse } from 'jest-mock-req-res';
-import { TypeResponse, sendGenericError, sendGenericSuccess } from '../../src/utils';
+import { TypeResponse, sendGenericError, sendGenericSuccess } from '../../utils';
 
 describe('Unit test sendGenericSuccess fn', () => {
   it('Happy path, return success response', () => {
